requirements.txt
fastapi
uvicorn
streamlit
openai
langchain
tiktoken
chromadb
python-dotenv
pydantic